Redid kv (again)
Had to redo these because using custom KV requests and responses for
different types of KV actions was difficult in Rust + would add a perf
hit.

The changes here now mean that KV host func will accept a *protos.KVStep
for the params and will return a *protos.KVStepResponse.

I also opted to share the usage of protos.KVAction between usage in
Steps and in KV instructions. This should help avoid confusion.