no need for any of the normalization stuff for Audience elements
If we allow spaces/non-[a-zA-Z0-9_-]+ for Audience elements, all SDKs
will have to know how to deal with such characters as well which would
be a huge pain in the ass. It is easier for us to enforce [a-zA-Z0-9-_]+
for now and avoid the whole space hassle.

The biggest pain was not allowing spaces for pipeline and step names and
that has been updated - so we should be good.