support for resuming tails + small logging improvement
Prior to the changes in these commits, there was no way to resume tails.
As in, if an active tail was taking place from the UI and the SDK client was
restarted - the tail output would not resume in the UI until you stopped
the previous Tail() request and started a new one.

This change introduces new functionality - the server will noiw send all
active tail commands when an SDK calls `internal.Register()`.

This means that the SDK may be given tail commands for an audience that
it has not seen yet. Because of this, the SDK will now record the tail
request and only start it when it sees the audience for the first time
(in a `sdk.Process()` call).

---

2nd part: modified "no-op" logger to "TinyLogger" -- the tiny logger
will now suppress Debug and Info messages, while still display Warning
and Error messages.