audiences done('ish)
Update needs to handle a delete of objects - currently does not.

Ie. If you apply a CR that defines two audiences, then the CR is updated
and one of the audiences is removed - that should be handled as a
partial delete.

Right now, deletes are only handled if there is an actual `kubectl
delete ...` only which is incorrect.